fix: trait-result build error
